Each file has 1 to 3 trivial casts added and all involve things whose known range ensures the change is ok (boolean, address-size, and other known things)..
Changes fix new-ish warnings from MSVC.
	modified:   src/bin/dwarfdump/dd_uri.c
	modified:   src/lib/libdwarf/dwarf_debugaddr.c
	modified:   src/lib/libdwarf/dwarf_generic_init.c
	modified:   src/lib/libdwarf/dwarf_gnu_index.c
	modified:   src/lib/libdwarf/dwarf_init_finish.c
	modified:   src/lib/libdwarf/dwarf_line.c
	modified:   src/lib/libdwarf/dwarf_line_table_reader_common.h
	modified:   src/lib/libdwarf/dwarf_loc.c
	modified:   src/lib/libdwarf/dwarf_setup_sections.c
	modified:   src/lib/libdwarfp/dwarf_pro_log_extra_flag_strings.c
	modified:   test/test_dwarfstring.c
	modified:   test/test_errmsglist.c
	modified:   test/test_linkedtopath.c

+++ test/test_dwarfstring.c
@@ -185,12 +185,15 @@ test1(void)
     d = dwarfstring_string(&g);
     check_string("expected bigstring ",bigstr,d,__LINE__);
     biglen = dwarfstring_strlen(&g);
-    check_value("expected 120  ",strlen(bigstr),biglen,__LINE__);
+    check_value("expected 120  ",(unsigned long)strlen(bigstr),
+        (unsigned long)biglen,__LINE__);
 
     dwarfstring_append_length(&g,"xxyyzz",3);
 
     biglen = dwarfstring_strlen(&g);
-    check_value("expected 123  ",strlen(bigstr)+3,biglen,__LINE__);
+    check_value("expected 123  ",
+        (unsigned long)strlen(bigstr)+3,
+        (unsigned long)biglen,__LINE__);
     dwarfstring_destructor(&g);
     return 0;
 }
@@ -233,7 +236,9 @@ test2(void)
     d = dwarfstring_string(&g);
     check_string("expected bigstring ",bigstr,d,__LINE__);
     biglen = dwarfstring_strlen(&g);
-    check_value("expected 120  ",strlen(bigstr),biglen,__LINE__);
+    check_value("expected 120  ",
+        (unsigned long)strlen(bigstr),
+        (unsigned long)biglen,__LINE__);
     dwarfstring_destructor(&g);
     return 0;
 }
@@ -300,8 +305,8 @@ test4(void)
     d = dwarfstring_string(&g);
     check_string("expected a01234xyz ",(char *)targetmystr,
         d,__LINE__);
-    check_value("expected 9",9,dwarfstring_strlen(&g),__LINE__);
-
+    check_value("expected 9",(unsigned long)9,
+        (unsigned long)dwarfstring_strlen(&g),__LINE__);
     if (d == fixedarea) {
         printf(" FAIL  not reallocated but should  be! line %d ",
             __LINE__);
